Compile project using Makefile (#456)
Coq compilation using `make`

Implement the compilation of the Coq part of the project using Makefile, without relying on the `configure.sh` script. With this approach it is needed to only run `make` in order to compile the project.

* `make CoqMakefile` implements the same thing as the `configure.sh` script did: executes `coq_makefile` to generate another Makefile to compile Coq
* `make Name.vo` can be used to compile an individual file, `Name.v`
* `make all` (which is also the default target) can be used to compile all Coq files in the project
* `make clean` can be used to delete all build artifacts

CI is also updated to be compatible with this compilation method.